+++ meta_package_manager/__main__.py
@@ -45,7 +45,7 @@ def main():
 
         - plain inline package call: ``python -m meta_package_manager``,
         - Poetry's script entry point: ``mpm = 'meta_package_manager.__main__:main``,
-        - Nuitka's main module invokation requirement: ``python -m nuitka (...)
+        - Nuitka's main module invocation requirement: ``python -m nuitka (...)
           meta_package_manager/__main__.py``
 
     That way we can deduce all three cases from the Poetry entry point.